moved ERR_ACTION_INVALID to general messages
