+++ internal/messages/messages.go
@@ -1,3 +1,4 @@
 package messages
 
+const ERR_ACTION_INVALID string = "invalid option selected"
 const ERR_RANDLEN_GEN string = "error generating the length: %s"
